+++ app/helpers/page_helper.rb
@@ -1,0 +1,5 @@
+module PageHelper
+  def page(identifier)
+    Page.find_in_cache(identifier).decorate
+  end
+end